Added getAll, get, getTags, edit and delete endpoints in controller
